working on communication

+++ src/monitor/filters.h
@@ -13,7 +13,7 @@ namespace monitor
 
 struct CMediumClassFilter : public common::CMonitorMediumFilter
 {
-	CMediumClassFilter( common::CMediumKinds::Enum _mediumClass, int _mediumNumber = -1 ):
+	CMediumClassFilter( common::CMediumKinds::Enum _mediumClass, int _mediumNumber = ( unsigned int )-1 ):
 		m_mediumClass( _mediumClass ),
 		m_mediumNumber( _mediumNumber )
 	{}
@@ -23,7 +23,7 @@ struct CMediumClassFilter : public common::CMonitorMediumFilter
 		std::list< common::CMonitorBaseMedium *> mediums;
 		mediums = _nodesManager->getNodesByClass( m_mediumClass );
 
-		if ( m_mediumNumber != -1 && mediums.size() > m_mediumNumber )
+		if ( m_mediumNumber != ( unsigned int )-1 && mediums.size() > m_mediumNumber )
 		{
 			mediums.resize( m_mediumNumber );
 		}
@@ -36,7 +36,7 @@ struct CMediumClassFilter : public common::CMonitorMediumFilter
 	}
 
 	common::CMediumKinds::Enum m_mediumClass;
-	int m_mediumNumber;
+	unsigned int m_mediumNumber;
 };
 
 struct CNodeExceptionFilter : public common::CMonitorMediumFilter
